SNOW-2183023: removed http_client for now
